Changes the print_border function to be 0 index
The long story short on this decision was the realization that k_printstr() is also 0 indexed. It was very silly to stubbornly attempt to stick to 1 index on the border print but not for main print. I decided to swap to this after getting *very* confused on the printing of that simple Run Process. This ate like an hour of my time due to a single bad decision.